refact: remoção de código não utilizado e cometários, criação do BaseChartComponent.js
