+++ dashboard/index.html
@@ -4,31 +4,25 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Dashboard Fitness</title>
+
+    <link href="../src/styles/bootstrap/bootstrap.css" rel="stylesheet">
+    <link href="../src/styles/main.css" rel="stylesheet">
     <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
-    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet">
 </head>
 <body>
 
-    <div class="container">
+<h1>Dashboard</h1>
 
-        <div class="rectangle-b565e209b699 shape grafico-treino">
-            <canvas id="line-chart" class="line-chart"></canvas>
-        </div>
+<div class="chart-container">
+    <h3>Treinos por Semana</h3>
+    <canvas id="training-per-weekly-chart"></canvas>
+</div>
 
-        <div style="height: 400px; width: 100%;">
-            <canvas id="training-per-weekly-chart"></canvas>
-        </div>
-
-        <div class="treino">
-            <button class="btn-rot" src="">Monte seu dashboard </button>
-        </div>
-    </div>
-        
-    
 </body>
 
 <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-<script src="index.js"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-ndDqU0Gzau9qJ1lfW4pNLlhNTkCfHzAVBReH9diLvGRem5+R9g2FzA8ZGN954O5Q"
+        crossorigin="anonymous"></script>
 <script src="../src/app.js" type="module"></script>
 </html>
