+++ dashboard/style.css
@@ -1,91 +1,4 @@
-html, body {
-  margin: 0;
-  min-height: 100%;
-  min-width: 100%;
-  padding: 0;
-}
-
-body {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-.text-node { background-clip: text !important; -webkit-background-clip: text !important; }
-
-/* Rectangle */
-.rectangle-b565e209b699 {
- 
-  left: 43px;
-  top: 20px;
-  width: 919px;
-  height: 378px;
-  background: #ffffffFF;
-  border: 3px solid #cfcfcfFF;
-  border-radius: 25px 25px 25px 25px;
-  box-shadow: 0px 0px 10px 0px #0000001A;
-}
-.shape {
-  display: flex;
-  width: 100%;
-  max-width: 600px; /* ajuste conforme necessário */
-  min-width: 200px;
-  margin: 0 auto;
-  box-sizing: border-box;
-}
-.line-chart {
-  width: 100% !important;
-  height: 100% !important;
-  display: flex;
-  padding: 5px;
-}
-.container{
-  display: block;
-}
-
-
-.btn-rot {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 50%;
-  max-width: 200px;
-  min-width: 120px;
-  height: auto;
-  min-height: 40px;
-  padding: 10px 18px;
-  border-radius: 20px;
-  border: none;
-  font-size: 0.8rem;
-  font-weight: 600;
-  background: linear-gradient(90deg, #5cfa1e 0%, #61f339 100%);
-  color: #fff;
-  box-shadow: 0 2px 8px 0 rgba(92, 250, 30, 0.15);
-  cursor: pointer;
-  transition: background 0.3s, transform 0.2s;
-  text-align: center;
-  word-break: break-word;
-  white-space: normal;
-}
-
-.btn-rot:hover {
-  background: linear-gradient(90deg, #61f339 0%, #5cfa1e 100%);
-  transform: translateY(-2px) scale(1.03);
-  box-shadow: 0 4px 16px 0 rgba(92, 250, 30, 0.25);
-}
-.treino{
-  display: flex;
-  margin-top: 5px;
-  margin-left: 5px;
-}
-
-body, button, input, select, textarea {
-  font-family: 'Poppins', Arial, Helvetica, sans-serif;
-  letter-spacing: 0.01em;
-  font-size: 16px;
-  color: #222;
+.chart-container {
+    height: 400px;
+    width: 100%;
 }
